% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{build_accordion}
\alias{build_accordion}
\title{Create a bootstrap accordion UI element}
\usage{
build_accordion(id, header, content)
}
\arguments{
\item{id}{(String) A unique id value for the accordion}

\item{header}{(String) The text to display in the header of the accordion
(to the left of the 'collapse' button)}

\item{content}{The content (HTML, text, etc.) to be displayed within the 
accordion}
}
\value{
An HTML div that will display `content` inside a vertically-
  collapsible bootstrap accordion
}
\description{
In bootstrap 5.0 you can create an "accordion" object that allows you to 
collapse/expand the content inside the "accordion" container:
\link{https://getbootstrap.com/docs/5.0/components/accordion/}
}
\examples{
if (interactive()) {
  library(shiny)
  library(bslib)

  ui <- page_fluid(
    build_accordion(
      id = "acc1", 
      header = "Open/Close Me", 
      content = tags$em("Here's some collapsible content")
    )
  )

  server <- function(input, output) {}

  shinyApp(ui, server)
}
}
