% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{lookup_element_name_by_value}
\alias{lookup_element_name_by_value}
\title{Look up the name of an element in a named character vector}
\usage{
lookup_element_name_by_value(x, value)
}
\arguments{
\item{x}{A named character vector}

\item{value}{The value in `x` to use to look up the corresponding name}
}
\value{
A character string representing the name of the value in the vector
  specified by `value`
}
\description{
This function was developed as an easy way to extract the corresponding name 
  from a named character vector given a value within that vector.
}
\examples{
# The following code would return "Tamarack"
lookup_element_name_by_value(
  x = ELEMENT_NAMES$tree,
  value = "tree-lari-lar"
)
}
