% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{lookup_element_type_by_value}
\alias{lookup_element_type_by_value}
\title{Look up the name of the list element based upon item in one list}
\usage{
lookup_element_type_by_value(x, value)
}
\arguments{
\item{x}{A list object}

\item{value}{The value in one of the lists of `x` to use to look up the 
name of the list element(s) containing `x`}
}
\value{
A character string representing the name(s) of the list(s) containing
  `x`
}
\description{
This function was developed as a pipe-friendly way to extract the 
  corresponding name from a named character vector given a value within that
  vector.
}
\examples{
# The following code would return "Tamarack"
lookup_element_type_by_value(
  x = ELEMENT_NAMES,
  value = "tree-lari-lar"
)
}
