% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBI-package.R
\docType{package}
\name{WBI-package}
\alias{WBI-package}
\title{WBI: Wasserstein Bipolarization Index.}
\description{
Computation of the Wasserstein Bipolarization Index as described in Lee and Sobel (Forthcoming).
}
\author{
\strong{Maintainer}: Hane Lee \email{hl3333@columbia.edu}

}
\keyword{internal}
