% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBI_boot.R
\name{WBI_boot}
\alias{WBI_boot}
\title{Wasserstein Bipolarization Index with Bootstrap Confidence Intervals}
\usage{
WBI_boot(obs, wp, min, max, alpha, r = 5000)
}
\arguments{
\item{obs}{A vector of observed responses.}

\item{wp}{The order (p) of the p-Wasserstein distance. Use p=1 to only account for spread, p>1 to account for spread and bi-clustering.}

\item{min}{The minimum of the response scale.}

\item{max}{The maximum of the response scale.}

\item{alpha}{Significance level of the confidence interval.}

\item{r}{Number of Bootstrap replications}
}
\value{
A vector of length three containing the CI lower bound, point estimate, and CI upper bound.
}
\description{
This function takes a vector of observed responses and computes the Wasserstein
Bipolarization Index and its bootstrap confidence interval with the maximum
polarization distribution (with 0.5 masses each on min and max) as the maximally
separated measure. We use the package bcaboot for bootstrap.
}
\examples{
# We want to measure WBI of an opinion item measured on [0,100].
# We choose W2 distance and 95\% bootstrap confidence intervals.
data <- c(20, 100, 50, 50, 0,
          90, 85, 10, 25, 10,
          30, 90, 80, 0, 100,
          20, 30, 0, 65, 95)
results <- WBI_boot(data, 2, 0, 100, 0.05)

}
